h-tests: add basic infrastructure to create dummy code-root setup
So far, I was hoping I could check-in some dummy code-root setups
directly to the git repository as fixtures. Sadly, it's almost
impossible to check-in git repos in a git repo. The only option do to
so would be to setup git subrepo. Needless to say, I *really* don't
want to go down that path.

Instead, I created some infrastructure to create a temp directory in
which we generate the test code-roots on the fly.