h--parse-repo-identifier: actually parse the query string
This function used to only return the kind of the query string. This
is not sufficient to properly query the forges. On top of returning
the type of a query string, we want this function to actually parse
the query into a tagged union materialized as a astring.

We update the tests accordingly to this refactoring.