h-jump-to-project: clone project if not available in the code root.
We introduce a new autocomplete function to help us determine whether
the user-query was already in the collection of the completing read
call or whether this is a user-specified value.

We then take advantage of this function to make a single entry point
for the whole plugin: h-jump-to-project. This function now not only
allow a user to jump to a local project but also clones a missing
project from the available forge sources.