h--call-git-in-dir: funcall the callback even in case of failure
We had a bug when git was exiting with a strange error code. We were
ending up with a event in the form of:

    exited abnormally with code 128

I'm not sure how to properly handle these events. I fixed this error
by calling the callback regardless of the event. It might backfire in
the future though: in some cases, a process can trigger a event
without exiting according to the manual. In that case, we'll end up
calling the callback with nil as exit code. I'm not 100% sure h.el
will handle that gracefully.

That being said, I never encountered this use case so far. I guess
it'll be a problem for future me to solve!